+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Release_1.3.0_20251121_build_A
 
 ### 功能构建
+  
+- 优化支持实体机制。
+  - 新建支持 QoS 服务 com.dwarfeng.statistics.stack.service.SupportQosService。
+  - 将支持实体维护服务的重置功能迁移至 QoS 服务。
 
 - 日志结构升级。
   - 修改 logging 目录下配置文件名称。